Cities and bugfix

+++ server_files/server_attackable.h
@@ -1,6 +1,8 @@
 #ifndef _SERVER_ATTACKABLE
 #define _SERVER_ATTACKABLE
 
+#include <stdint.h>
+
 /* 
  * Clase que representa a un atacable del juego.
  * Va a funcionar como interfaz para poder hacer
@@ -15,10 +17,10 @@ public:
     virtual bool is_safe() = 0;
     virtual bool is_newbie() = 0;
     virtual bool is_alive() = 0;
-    virtual int get_max_life() = 0;
-    virtual int get_level() = 0;
-    virtual int get_body_pos_X() const = 0;
-    virtual int get_body_pos_Y() const = 0;
+    virtual int16_t get_max_life() = 0;
+    virtual int16_t get_level() = 0;
+    virtual int16_t get_body_pos_X() const = 0;
+    virtual int16_t get_body_pos_Y() const = 0;
     virtual int defense(int damage) = 0;
     virtual void attack(Attackable& other) = 0;
 };
