+++ server_files/server_character_class.h
@@ -7,7 +7,7 @@
 /* Clase que representa a una clase del juego. Posee algunos stats. */
 class CharacterClass {
 protected:
-    int16_t id;
+    uint8_t id;
     float life_multiplier;
     float mana_multiplier;
     float meditation_multiplier;
@@ -17,7 +17,7 @@ public:
     CharacterClass();
 
     // Devuelve el id de la clase
-    int16_t get_id();
+    uint8_t get_id();
 
     // Devuelve el multiplicador de vida de la clase
     float get_life_multiplier();
